requirements.txt
streamlit
requests
accelerate
pandas
psycopg2-binary
flask 
transformers 
sqlalchemy
torch
bitsandbytes